# refactor: Improve db sync performance