Replaces the two document caches with a single events cache in gatekeeper. An augmented `admin perf-test` measured a 300X performance increase for DID exports (e.g. from 65 seconds to 146ms for 1440 DIDs on one node). This refactor essentially mitigates the load jump seen in #255. 

Even though we are no longer caching documents, and have to reconstruct them from the events cache, no significant change was measured for the resolution benchmarks.

Incidentally it is possible to measure the size of the events cache by noting the number of bytes returned by a call to /export-dids. In this case ~4.6 MB for 1446 DIDs:

```
kc-gatekeeper-1  | POST /api/v1/export-dids 200 90.996 ms - 4585370
```

The response time of 91ms translates to ~63 _microseconds_ per DID.

This PR also adds unit tests for generateDoc in order to pass the coverage check. (It seems that just removing code can have the side effect of reducing coverage% which fails the check).